+++ src/main/java/wayoftime/bloodmagic/api/compat/IAltarReader.java
@@ -1,4 +1,4 @@
-package wayoftime.bloodmagic.api.item;
+package wayoftime.bloodmagic.api.compat;
 
 /**
  * Any item that implements this interface will not be pulled into the Altar on
