+++ src/main/java/wayoftime/bloodmagic/api/compat/IDemonWillWeapon.java
@@ -1,4 +1,4 @@
-package wayoftime.bloodmagic.api.will;
+package wayoftime.bloodmagic.api.compat;
 
 import java.util.List;
