+++ src/main/java/wayoftime/bloodmagic/client/hud/element/ElementDemonAura.java
@@ -12,7 +12,7 @@ import net.minecraft.util.ResourceLocation;
 import wayoftime.bloodmagic.BloodMagic;
 import wayoftime.bloodmagic.util.Utils;
 import wayoftime.bloodmagic.util.handler.event.ClientHandler;
-import wayoftime.bloodmagic.api.will.EnumDemonWillType;
+import wayoftime.bloodmagic.api.compat.EnumDemonWillType;
 
 public class ElementDemonAura extends HUDElement
 {
