+++ src/main/java/wayoftime/bloodmagic/common/block/BlockPath.java
@@ -4,7 +4,7 @@ import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import wayoftime.bloodmagic.api.incense.IIncensePath;
+import wayoftime.bloodmagic.api.compat.IIncensePath;
 
 public class BlockPath extends Block implements IIncensePath
 {
