+++ src/main/java/wayoftime/bloodmagic/common/item/ItemDemonCrystal.java
@@ -3,8 +3,8 @@ package wayoftime.bloodmagic.common.item;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import wayoftime.bloodmagic.BloodMagic;
-import wayoftime.bloodmagic.api.will.EnumDemonWillType;
-import wayoftime.bloodmagic.api.will.IDiscreteDemonWill;
+import wayoftime.bloodmagic.api.compat.EnumDemonWillType;
+import wayoftime.bloodmagic.api.compat.IDiscreteDemonWill;
 
 public class ItemDemonCrystal extends Item implements IDiscreteDemonWill
 {
