+++ src/main/java/wayoftime/bloodmagic/common/item/ItemDemonWillGauge.java
@@ -13,7 +13,7 @@ import net.minecraft.world.World;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import wayoftime.bloodmagic.BloodMagic;
-import wayoftime.bloodmagic.api.will.IDemonWillViewer;
+import wayoftime.bloodmagic.api.compat.IDemonWillViewer;
 import wayoftime.bloodmagic.util.handler.event.GenericHandler;
 
 public class ItemDemonWillGauge extends Item implements IDemonWillViewer
