+++ src/main/java/wayoftime/bloodmagic/common/item/soul/ItemSentientSword.java
@@ -35,12 +35,12 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 import wayoftime.bloodmagic.BloodMagic;
 import wayoftime.bloodmagic.common.item.BMItemTier;
 import wayoftime.bloodmagic.common.item.BloodMagicItems;
-import wayoftime.bloodmagic.api.will.IMultiWillTool;
+import wayoftime.bloodmagic.api.compat.IMultiWillTool;
 import wayoftime.bloodmagic.util.Constants;
 import wayoftime.bloodmagic.util.helper.NBTHelper;
-import wayoftime.bloodmagic.api.will.EnumDemonWillType;
-import wayoftime.bloodmagic.api.will.IDemonWill;
-import wayoftime.bloodmagic.api.will.IDemonWillWeapon;
+import wayoftime.bloodmagic.api.compat.EnumDemonWillType;
+import wayoftime.bloodmagic.api.compat.IDemonWill;
+import wayoftime.bloodmagic.api.compat.IDemonWillWeapon;
 import wayoftime.bloodmagic.will.PlayerDemonWillHandler;
 
 public class ItemSentientSword extends SwordItem implements IDemonWillWeapon, IMultiWillTool
