+++ src/main/java/wayoftime/bloodmagic/demonaura/WorldDemonWillHandler.java
@@ -12,7 +12,7 @@ import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.chunk.IChunk;
 import wayoftime.bloodmagic.util.BMLog;
 import wayoftime.bloodmagic.will.DemonWillHolder;
-import wayoftime.bloodmagic.api.will.EnumDemonWillType;
+import wayoftime.bloodmagic.api.compat.EnumDemonWillType;
 
 public class WorldDemonWillHandler
 {
