+++ src/main/java/wayoftime/bloodmagic/tile/TileIncenseAltar.java
@@ -18,7 +18,7 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.server.ServerWorld;
 import net.minecraftforge.registries.ObjectHolder;
 import wayoftime.bloodmagic.incense.EnumTranquilityType;
-import wayoftime.bloodmagic.api.incense.IIncensePath;
+import wayoftime.bloodmagic.api.compat.IIncensePath;
 import wayoftime.bloodmagic.incense.IncenseAltarHandler;
 import wayoftime.bloodmagic.incense.IncenseTranquilityRegistry;
 import wayoftime.bloodmagic.incense.TranquilityStack;
